move print_command to command.c
